extract search group item and apply to public profile
