+++ frontend/src/pages/user/PublicProfilePage.jsx
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from "react";
-import  Button  from "../../components/Button";
+import Button from "../../components/Button";
 import { useParams } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { IoMdMale, IoMdFemale } from "react-icons/io";
 import { useAuth } from "../../store/AuthContext";
 import UserGroupBar from "../../components/UserGroupBar";
+import SingleSearchedGroup from "../../components/SingleSearchedGroup";
 
 const PublicProfilePage = () => {
   const { isLoggedIn, user: loggedInUser, setIsLoggedIn } = useAuth();
@@ -20,11 +21,16 @@ const PublicProfilePage = () => {
     const fetchUserData = async () => {
       // Fetch user data from API, if the user is not logged in or viewing another user's profile
       try {
-
-        let { data } = await axios.get(`${import.meta.env.VITE_API_BASE_URL}/api/user/userData/${userId}`);
+        let { data } = await axios.get(
+          `${import.meta.env.VITE_API_BASE_URL}/api/user/userData/${userId}`
+        );
         setUser(data);
 
-        if (data && data.participatingGroups && data.participatingGroups.length > 0) {
+        if (
+          data &&
+          data.participatingGroups &&
+          data.participatingGroups.length > 0
+        ) {
           // sort the groups by time
           data.participatingGroups.sort((a, b) => {
             return new Date(b.createDate) - new Date(a.createDate);
@@ -42,16 +48,23 @@ const PublicProfilePage = () => {
 
       setUser(loggedInUser);
       setGroups(loggedInUser.participatingGroups);
+      console.log(loggedInUser.participatingGroups);
       setIsLoading(false);
     } else {
-
       fetchUserData();
     }
   }, [userId, isLoggedIn, loggedInUser]);
 
   if (isLoading) {
     return (
-      <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
+      <div
+        style={{
+          display: "flex",
+          justifyContent: "center",
+          alignItems: "center",
+          height: "100vh",
+        }}
+      >
         <img src="/image/Spinner.svg" alt="Loading..." />
       </div>
     );
@@ -66,15 +79,22 @@ const PublicProfilePage = () => {
       <div className="m-4 p-4  bg-white flex flex-col flex-grow">
         <div className="py-8">
           {/* Add new feature for tags reminder */}
-        {isLoggedIn && userId === loggedInUser._id && (!user.profileTags || user.profileTags.length === 0) && (
-          <div className="fixed w-[280px] right-10 bottom-10 p-4 bg-blue-100 rounded-lg shadow-lg">
-            <p className="">It seems you haven't added any tags in your profile! Tag yourself can make people find you easily!</p>
-            <Button className="mt-2"
-              onClick={() => navigate(`/user/settings/${user._id}`)}>
-              Go to Settings
-            </Button>
-          </div>
-        )}
+          {isLoggedIn &&
+            userId === loggedInUser._id &&
+            (!user.profileTags || user.profileTags.length === 0) && (
+              <div className="fixed w-[280px] right-10 bottom-10 p-4 bg-blue-100 rounded-lg shadow-lg">
+                <p className="">
+                  It seems you haven't added any tags in your profile! Tag
+                  yourself can make people find you easily!
+                </p>
+                <Button
+                  className="mt-2"
+                  onClick={() => navigate(`/user/settings/${user._id}`)}
+                >
+                  Go to Settings
+                </Button>
+              </div>
+            )}
           <div className="flex items-center mb-2">
             <img
               className="w-40 h-40 rounded-full mr-4"
@@ -91,38 +111,46 @@ const PublicProfilePage = () => {
                   <IoMdFemale className="fill-pink-500 text-2xl mt-5" />
                 )}
                 {user.isVerification ? (
-                  <div className="p-1 px-2 rounded-full text-center bg-primary text-white text-xs">Verified</div>
+                  <div className="p-1 px-2 rounded-full text-center bg-primary text-white text-xs">
+                    Verified
+                  </div>
                 ) : (
-                  <div className="p-1 px-2  rounded-full text-center bg-gray-400 text-white text-xs">Unverified</div>
+                  <div className="p-1 px-2  rounded-full text-center bg-gray-400 text-white text-xs">
+                    Unverified
+                  </div>
                 )}
               </div>
 
               <div className="ml-5 flex items-center ">Email: {user.email}</div>
               <div className="flex ml-5 mb-5 flex-wrap">
-                {user.profileTags && user.profileTags.map((tag) => (
-                  <ProfileTags key={tag._id} tagName={tag.name} />
-                ))}
+                {user.profileTags &&
+                  user.profileTags.map((tag) => (
+                    <ProfileTags key={tag._id} tagName={tag.name} />
+                  ))}
               </div>
-              {isLoggedIn && userId === loggedInUser._id && !user.isVerification && (
-                <div className="ml-3 border-2 rounded-full py-1 bg-amber-200 text-xs px-3 text-gray-600">
-                  Your account is not verified. Unverified accounts may not create a group in HeyMate.
-                 Please go to Settings.</div>
-              )}
-
+              {isLoggedIn &&
+                userId === loggedInUser._id &&
+                !user.isVerification && (
+                  <div className="ml-3 border-2 rounded-full py-1 bg-amber-200 text-xs px-3 text-gray-600">
+                    Your account is not verified. Unverified accounts may not
+                    create a group in HeyMate. Please go to Settings.
+                  </div>
+                )}
             </div>
-
           </div>
         </div>
         {/* Groups section */}
         <div className="text-3xl mb-8">Groups</div>
         <div className="">
           {isLoggedIn ? (
-            groups && groups.length > 0 ? groups.map((group) => (
-              <UserGroupBar key={group._id} group={group} />
-            )) : (
+            groups && groups.length > 0 ? (
+              groups.map((group) => (
+                // <UserGroupBar key={group._id} group={group} />
+                <SingleSearchedGroup key={group._id} group={group} />
+              ))
+            ) : (
               <p>No groups found.</p>
             )
-
           ) : (
             <div className="text-lg text-gray-400 ">Please log in first.</div>
           )}
